use rgba16f for color

+++ shaders/lib/buffers.glsl
@@ -1,5 +1,5 @@
 /*
-    const int colortex0Format = RGBA32F;
+    const int colortex0Format = RGBA16F;
     const int colortex1Format = RGBA32F;
     const int colortex2Format = RGBA32F;
     const int colortex7Format = RGBA32F;
